On branch master  Changes to be committed: 	modified:   Makefile 	modified:   src/action.c 	modified:   src/action.h 	modified:   src/main.c 	modified:   src/parse.c
 Better organized find_class:
	 * Moved all pointer declerations to the beginning
	 * Added a free for all pointers at the end
	 * ^ Closed potential memory leak issue with too many pointer declerations if the main loop of
	   find_class was too long because data/time.txt was longer than what I test with.

+++ src/parse.c
@@ -24,7 +24,7 @@ Assumptions:
 	Ints n and m are nonnegative, within range of linebuf, and m is greater than n
 */
 int parse_class_time(char *linebuf, int n, int m) {
-	if(linebuf == EOF || linebuf == 0 || n < 0 || m < n) return EOF;
+	if((long)linebuf == EOF || (long)linebuf == 0 || n < 0 || m < n) return EOF;
 
 	if(vbose > 1) { fprintf(stdout, "Parsing line '%s' for time between indices '%d' and '%d'\n", linebuf, n, m); }
